Tela, Carta, Funcionalidades de Gestao

+++ db/migrate/20170608041910_create_pokers.rb
@@ -2,7 +2,7 @@ class CreatePokers < ActiveRecord::Migration
   def change
     create_table :pokers do |t|
       t.string :name
-      t.integer :value
+      t.decimal :value
 
       t.timestamps null: false
     end
